requirements.txt
asgiref==3.5.2
build==0.8.0
click==8.1.3
Django==4.1.1
django-crispy-forms==1.14.0
djangorestframework==3.14.0
packaging==21.3
pep517==0.13.0
pip-tools==6.8.0
psycopg2-binary==2.9.3
pyparsing==3.0.9
pytz==2022.4
sqlparse==0.4.2
tomli==2.0.1
